requirements.txt
BeautifulSoup==3.2.1
Django==1.6.1
MySQL-python==1.2.5
Pillow==2.3.0
PyYAML==3.10
cssselect==0.9.1
goose-extractor==1.0.6
jieba==0.31
lxml==3.3.0
nltk==2.0.4
requests==2.2.1
wsgiref==0.1.2
